+++ libs/store/src/__tests__/store.test.ts
@@ -1,7 +1,11 @@
 import 'reflect-metadata';
-import { resolve, Instantiable, container, unionResolve } from '../lib/store';
 import { Service1, Service2, Store2 } from './util';
-
+import {
+  resolve,
+  Instantiable,
+  unionResolve,
+  container,
+} from '../lib/defaultContainer';
 test('Resolve should work', () => {
   const instance = resolve(Service1);
   expect(instance instanceof Service1).toBe(true);
