+++ libs/store/src/lib/constants.ts
@@ -1,0 +1,3 @@
+export const CLASS_SYMBOL = Symbol('classSymbol');
+export const SINGLETON_SYMBOL = Symbol('isSingleton');
+export const USE_SYMBOL_AS_IDENTIFIER = false;
