+++ libs/store/src/lib/types.ts
@@ -9,3 +9,9 @@ export class BaseController {
 }
 
 export type StoreContextType = React.Context<Container>;
+
+
+export type InjectableConfig = {
+  name?: string;
+  singleton?: boolean;
+};
